New displaytype +bug fix
Added new shop displayType. And will piggyback off of that for the enchanted books for the names but that is not implemented yet
Forget when I put a fix in for this. But it is shopbugs/issues/87
The shop check from earlier commit was broken should be fixed
Starts on work on making sure doubles can be worked on currency